+++ src/main/java/com/kai/mynote/controller/WorkSpaceController.java
@@ -8,12 +8,15 @@ import com.kai.mynote.entities.WorkSpace;
 import com.kai.mynote.service.Impl.UserServiceImpl;
 import com.kai.mynote.service.Impl.WorkspaceServiceImpl;
 import com.sun.tools.jconsole.JConsoleContext;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+
+import java.io.IOException;
 
 @RestController
 @RequestMapping("/workspace")
@@ -36,7 +39,7 @@ public class WorkSpaceController {
     @GetMapping("/{id}")
     public Page<Note> getNotesByWorkspaceId(@PathVariable Long id, Authentication authentication,
                                             @RequestParam(defaultValue = "0") int page,
-                                            @RequestParam(defaultValue = "30") int size){
+                                            @RequestParam(defaultValue = "30") int size) {
         WorkSpace workSpace = workspaceService.getWorkspaceById(id);
         if (authentication.getName().equalsIgnoreCase(workSpace.getAuthor().getUsername())){
             return workspaceService.getAllNoteByWorkspaceId(id, page, size);
@@ -74,17 +77,17 @@ public class WorkSpaceController {
     @DeleteMapping("/remove/{id}")
     public ResponseEntity<ResponseObject> deleteWs(@PathVariable Long id, Authentication authentication){
         WorkSpace ws = workspaceService.getWorkspaceById(id);
-        if (ws != null && ws.getNotes().size() == 0){
-            if (ws.getAuthor().getUsername().equalsIgnoreCase(authentication.getName())){
+        if (ws != null && ws.getAuthor().getUsername().equalsIgnoreCase(authentication.getName())){
+            if (ws.getNotes().size() == 0){
                 workspaceService.removeById(id);
                 return ResponseEntity.status(HttpStatus.OK).body(
                         new ResponseObject(AppConstants.SUCCESS_STATUS,AppConstants.WORKSPACE +" "+AppConstants.REMOVED, null)
                 );
+            }else {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
+                        new ResponseObject(AppConstants.FAILURE_STATUS,AppConstants.WORKSPACE+" have to empty to remove",null)
+                );
             }
-        }else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
-                    new ResponseObject(AppConstants.FAILURE_STATUS,AppConstants.WORKSPACE+" have to empty to remove",null)
-            );
         }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
                 new ResponseObject(AppConstants.FAILURE_STATUS,"Bad request",null)
